src/scraper/requirements.txt
selenium
enum34
python-dotenv
watchdog[watchmedo]
kafka-python